+++ example/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:layrz_theme/layrz_theme.dart';
 import 'package:layrz_theme_example/router.dart';
-import 'package:layrz_theme_example/store.dart';
+import 'package:layrz_theme_example/store/store.dart';
 import 'package:vxstate/vxstate.dart';
 
 void main() {
@@ -24,7 +24,7 @@ class _MyAppState extends State<MyApp> {
   @override
   Widget build(BuildContext context) {
     return VxBuilder<AppStore>(
-      mutations: const {ToggleTheme},
+      mutations: const {SetTheme},
       builder: (context, store, status) {
         return MaterialApp.router(
           title: 'Layrz Theme Example',